captionmodel/imagecaption/requirements.txt
absl-py==1.1.0
appdirs==1.4.4
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
backports.print-function==1.1.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.2
executing==0.10.0
filelock==3.0.12
Flask==2.1.3
flatbuffers==1.12
fonttools==4.33.3
ftfy==6.1.1
future==0.18.2
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gTTS==2.2.4
gunicorn==20.1.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
imageio==2.21.1
importlib-metadata==4.11.4
imutils==0.5.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mpmath==1.2.1
networkx==2.8.5
numpy==1.23.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pad-sequences==0.6.1
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.30
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.7.25
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
stack-data==0.4.0
sympy==1.10.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.8.12
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.1
transforms==0.1
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
